--- src/webview/const.ts
@@ -1,10 +1,0 @@
-export const WELCOME_MESSAGE = `
-Hi! I'm twinny your pair programming buddy! \nType a message and press enter to send.
-
-If you highlight code in your editor I will be able to help you with things like refactoring and explaining.
-
-`
-
-export const BOT_NAME = 'twinny'
-
-export const USER_NAME = 'user'
