+++ src/webview/constants.ts
@@ -1,0 +1,3 @@
+export const BOT_NAME = 'twinny'
+
+export const USER_NAME = 'user'
